# apps/web/src/api/generated/openapi-v2.json
{
  "components": {
    "schemas": {
...
      "ArtifactEntry": {
        "properties": {
          "kind": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "kind"
        ],
        "type": "object"
      },
...
      "FieldCatalog": {
        "properties": {
          "domain_generation_id": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "quantities": {
            "items": {
              "$ref": "#/components/schemas/FieldDescriptor"
            },
            "type": "array"
          },
          "revision": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "domain_generation_id",
          "quantities"
        ],
        "type": "object"
      },
...
      "SlicePlane": {
        "enum": [
          "xy",
          "xz",
          "yz"
        ],
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Professional session-scoped API for Fullmag platform, simulation control, data, visualization, and diagnostics.",
    "license": {
      "name": ""
    },
    "title": "Fullmag API v2",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/sessions": {
      "get": {
        "operationId": "sessions_get_sessions",
        "responses": {
          "200": {
            "description": "Available sessions"
          }
        },
        "tags": [
          "sessions"
        ]
      },
      "post": {
        "operationId": "sessions_post_sessions",
        "responses": {
          "200": {
            "description": "Current session returned"
          },
          "400": {
            "description": "The local runtime only exposes the current session"
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/v2/sessions/current": {
      "get": {
        "operationId": "sessions_get_sessions_current",
        "responses": {
          "200": {
            "description": "Current session summary"
          }
        },
        "tags": [
          "sessions"
        ]
      },
      "patch": {
        "operationId": "sessions_patch_sessions_current",
        "responses": {
          "200": {
            "description": "Current session summary"
          },
          "400": {
            "description": "Session mutation is not yet supported"
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/v2/sessions/current/analysis/eigen/branches.v2": {
      "get": {
        "operationId": "analysis_get_sessions_current_analysis_eigen_branches_v2",
        "responses": {
          "200": {
            "description": "Tracked eigen branches v2"
          },
          "404": {
            "description": "No branches v2 data"
          }
        },
        "tags": [
          "analysis"
        ]
      }
    },
    "/v2/sessions/current/analysis/eigen/dispersion.csv": {
      "get": {
        "operationId": "analysis_get_sessions_current_analysis_eigen_dispersion_csv",
        "responses": {
          "200": {
            "description": "Eigen dispersion CSV v2"
          },
          "404": {
            "description": "No dispersion CSV"
          }
        },
        "tags": [
          "analysis"
        ]
      }
    },
    "/v2/sessions/current/analysis/eigen/modes/{sample_index}/{mode_index}": {
      "get": {
        "operationId": "analysis_get_sessions_current_analysis_eigen_modes_sample_index_mode_index",
        "parameters": [
          {
            "description": "K-path sample index",
            "in": "path",
            "name": "sample_index",
            "required": true,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Raw mode index within the sample",
            "in": "path",
            "name": "mode_index",
            "required": true,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eigen mode artifact v2"
          },
          "404": {
            "description": "Mode not found"
          }
        },
        "tags": [
          "analysis"
        ]
      }
    },
    "/v2/sessions/current/analysis/eigen/spectrum.v2": {
      "get": {
        "operationId": "analysis_get_sessions_current_analysis_eigen_spectrum_v2",
        "responses": {
          "200": {
            "description": "Eigen spectrum artifact v2"
          },
          "404": {
            "description": "No eigen spectrum v2 artifact"
          }
        },
        "tags": [
          "analysis"
        ]
      }
    },
    "/v2/sessions/current/analysis/eigenmodes/branches": {
      "get": {
        "operationId": "analysis_get_sessions_current_analysis_eigenmodes_branches",
        "responses": {
          "200": {
            "description": "Tracked branches"
          },
          "404": {
            "description": "No branches data"
          }
        },
        "tags": [
          "analysis"
        ]
      }
    },
    "/v2/sessions/current/analysis/eigenmodes/dispersion": {
      "get": {
        "operationId": "analysis_get_sessions_current_analysis_eigenmodes_dispersion",
        "responses": {
          "200": {
            "description": "Eigen dispersion data"
          },
          "404": {
            "description": "No dispersion data"
          }
        },
        "tags": [
          "analysis"
        ]
      }
    },
    "/v2/sessions/current/analysis/eigenmodes/modes/{mode_id}": {
      "get": {
        "operationId": "analysis_get_sessions_current_analysis_eigenmodes_modes_mode_id",
        "parameters": [
          {
            "description": "Mode index",
            "in": "query",
            "name": "index",
            "required": true,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Optional k-sample index",
            "in": "query",
            "name": "sample_index",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eigen mode data"
          },
          "404": {
            "description": "Mode not found"
          }
        },
        "tags": [
          "analysis"
        ]
      }
    },
    "/v2/sessions/current/analysis/eigenmodes/spectrum": {
      "get": {
        "operationId": "analysis_get_sessions_current_analysis_eigenmodes_spectrum",
        "responses": {
          "200": {
            "description": "Eigen spectrum"
          },
          "404": {
            "description": "No eigen spectrum artifact"
          }
        },
        "tags": [
          "analysis"
        ]
      }
    },
    "/v2/sessions/current/data/artifacts": {
      "get": {
        "operationId": "data_get_sessions_current_data_artifacts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactEntry"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of artifacts"
          },
          "304": {
            "description": "Artifact list not modified for the supplied ETag"
          },
          "404": {
            "description": "No active workspace"
          }
        },
        "tags": [
          "data"
        ]
      }
    },
    "/v2/sessions/current/data/artifacts/{artifact_id}": {
      "get": {
        "operationId": "data_get_sessions_current_data_artifacts_artifact_id",
        "parameters": [
          {
            "description": "Artifact relative path",
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artifact content"
          },
          "404": {
            "description": "Artifact not found"
          }
        },
        "tags": [
          "data"
        ]
      }
    },
    "/v2/sessions/current/data/domain/meta": {
      "get": {
        "operationId": "data_get_sessions_current_data_domain_meta",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainMeta"
                }
              }
            },
            "description": "Domain metadata"
          },
          "404": {
            "description": "No active workspace"
          }
        },
        "tags": [
          "data"
        ]
      }
    },
    "/v2/sessions/current/data/domain/topology": {
      "get": {
        "operationId": "data_get_sessions_current_data_domain_topology",
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {}
            },
            "description": "Binary FEM topology (FMMT)"
          },
          "204": {
            "description": "Not applicable (FDM)"
          },
          "304": {
            "description": "Domain topology not modified for the supplied ETag"
          },
          "404": {
            "description": "No active workspace"
          }
        },
        "tags": [
          "data"
        ]
      }
    },
    "/v2/sessions/current/data/fields": {
      "get": {
        "operationId": "data_get_sessions_current_data_fields",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldCatalog"
                }
              }
            },
            "description": "Field catalog"
          },
          "404": {
            "description": "No active workspace"
          }
        },
        "tags": [
          "data"
        ]
      }
    },
    "/v2/sessions/current/data/fields/{quantity_id}/meta": {
      "get": {
        "operationId": "data_get_sessions_current_data_fields_quantity_id_meta",
        "parameters": [
          {
            "description": "Quantity identifier",
            "in": "path",
            "name": "quantity_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldMeta"
                }
              }
            },
            "description": "Field metadata"
          },
          "404": {
            "description": "Field not found"
          }
        },
        "tags": [
          "data"
        ]
      }
    },
...
    "/v2/sessions/current/data/fields/{quantity_id}/projection/profile": {
      "get": {
        "operationId": "data_get_sessions_current_data_fields_quantity_id_projection_profile",
        "parameters": [
          {
            "description": "Quantity identifier",
            "in": "path",
            "name": "quantity_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plane",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SlicePlane"
            }
          },
          {
            "description": "Component selection. Defaults to magnitude for vector fields.",
            "in": "query",
            "name": "component",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Full projection grid width.",
            "in": "query",
            "name": "x_size",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Full projection grid height.",
            "in": "query",
            "name": "y_size",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Pixel column to probe.",
            "in": "query",
            "name": "pixel_x",
            "required": true,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Pixel row to probe.",
            "in": "query",
            "name": "pixel_y",
            "required": true,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Maximum profile samples returned after depth sorting.",
            "in": "query",
            "name": "max_samples",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldProjectionProfile"
                }
              }
            },
            "description": "Depth-resolved FEM projection profile for one raster pixel"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "404": {
            "description": "Field not found"
          },
          "409": {
            "description": "Projection profile requires nodal FEM field and tetrahedral mesh"
          }
        },
        "tags": [
          "data"
        ]
      }
    },
...
    "/v2/sessions/current/diagnostics/engine-log": {
      "get": {
        "operationId": "diagnostics_get_sessions_current_diagnostics_engine_log",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineLogResource"
                }
              }
            },
            "description": "Engine log entries"
          },
          "304": {
            "description": "Engine log not modified for the supplied ETag"
          },
          "404": {
            "description": "No active workspace"
          }
        },
        "tags": [
          "diagnostics"
        ]
      }
    },
...
    "/v2/sessions/current/meshing/capabilities": {
      "get": {
        "operationId": "meshing_get_sessions_current_meshing_capabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeshCapabilitiesResource"
                }
              }
            },
            "description": "Current mesh capability read-model"
          },
          "404": {
            "description": "No active workspace or mesh summary"
          }
        },
        "tags": [
          "meshing"
        ]
      }
    },
...
  },
...
}